CRUD LPB

+++ database/migrations/2025_02_21_010724_create_l_p_b_details_table.php
@@ -15,10 +15,10 @@ return new class extends Migration
             $table->bigIncrements('id');
             $table->bigInteger('lpb_id');
             $table->string("product_code");
-            $table->integer("length");
-            $table->integer("diameter");
-            $table->integer("price");
-            $table->integer("qty");
+            $table->bigInteger("length");
+            $table->bigInteger("diameter");
+            $table->bigInteger("price");
+            $table->bigInteger("qty");
             $table->enum("quality", ["Afkir", "Super"]);
             $table->timestamps();
         });
